+++ example/lib/pages/overriden_duration_page.dart
@@ -1,0 +1,61 @@
+import 'package:better_player/better_player.dart';
+import 'package:better_player_example/constants.dart';
+import 'package:flutter/material.dart';
+
+class OverriddenDurationPage extends StatefulWidget {
+  @override
+  _OverriddenDurationPageState createState() => _OverriddenDurationPageState();
+}
+
+class _OverriddenDurationPageState extends State<OverriddenDurationPage> {
+  BetterPlayerController _betterPlayerController;
+
+  @override
+  void initState() {
+    BetterPlayerConfiguration betterPlayerConfiguration =
+        BetterPlayerConfiguration(
+      aspectRatio: 16 / 9,
+      fit: BoxFit.contain,
+      handleLifecycle: true,
+    );
+    _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
+    _setupDataSource();
+    super.initState();
+  }
+
+  void _setupDataSource() async {
+    BetterPlayerDataSource dataSource = BetterPlayerDataSource(
+      BetterPlayerDataSourceType.network,
+      Constants.elephantDreamVideoUrl,
+
+      ///Play only 10 seconds of this video.
+      overriddenDuration: Duration(seconds: 10),
+    );
+    _betterPlayerController.setupDataSource(dataSource);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Notification player"),
+      ),
+      body: Column(
+        children: [
+          const SizedBox(height: 8),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: Text(
+              "Click play on player to show notification in status bar.",
+              style: TextStyle(fontSize: 16),
+            ),
+          ),
+          AspectRatio(
+            aspectRatio: 16 / 9,
+            child: BetterPlayer(controller: _betterPlayerController),
+          ),
+        ],
+      ),
+    );
+  }
+}