+++ example/lib/pages/welcome_page.dart
@@ -15,6 +15,8 @@ import 'package:better_player_example/pages/memory_player_page.dart';
 import 'package:better_player_example/pages/normal_player_page.dart';
 import 'package:better_player_example/pages/notification_player_page.dart';
 import 'package:better_player_example/pages/overridden_aspect_ratio_page.dart';
+import 'package:better_player_example/pages/overriden_duration_page.dart';
+import 'package:better_player_example/pages/placeholder_until_play_page.dart';
 import 'package:better_player_example/pages/playlist_page.dart';
 import 'package:better_player_example/pages/resolutions_page.dart';
 import 'package:better_player_example/pages/reusable_video_list/reusable_video_list_page.dart';
@@ -129,8 +131,14 @@ class _WelcomePageState extends State<WelcomePage> {
       _buildExampleElementWidget("Fade placeholder page", () {
         _navigateToPage(FadePlaceholderPage());
       }),
-      _buildExampleElementWidget("Change player theme", () {
+      _buildExampleElementWidget("Placeholder until play page", () {
+        _navigateToPage(PlaceholderUntilPlayPage());
+      }),
+      _buildExampleElementWidget("Change player theme page", () {
         _navigateToPage(ChangePlayerThemePage());
+      }),
+      _buildExampleElementWidget("Overridden duration page", () {
+        _navigateToPage(OverriddenDurationPage());
       }),
     ];
   }