+++ ios/Classes/FLTBetterPlayerPlugin.m
@@ -97,11 +97,6 @@ static void* playbackBufferFullContext = &playbackBufferFullContext;
               options:0
               context:playbackBufferFullContext];
     
-    // Add an observer that will respond to itemDidPlayToEndTime
-    [[NSNotificationCenter defaultCenter] addObserver:self
-                                             selector:@selector(itemDidPlayToEndTime:)
-                                                 name:AVPlayerItemDidPlayToEndTimeNotification
-                                               object:item];
 }
 
 - (void)removeVideoOutput {
@@ -146,9 +141,12 @@ static void* playbackBufferFullContext = &playbackBufferFullContext;
     [[_player currentItem] removeObserver:self
                                forKeyPath:@"playbackBufferFull"
                                   context:playbackBufferFullContext];
+    [[NSNotificationCenter defaultCenter] removeObserver:self name:AVPlayerItemPlaybackStalledNotification object:nil];
     [[NSNotificationCenter defaultCenter] removeObserver:self];
     AVAsset* asset = [_player.currentItem asset];
     [asset cancelLoading];
+    
+    
 }
 
 - (void)itemDidPlayToEndTime:(NSNotification*)notification {
@@ -251,12 +249,12 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
     return transform;
 }
 
-- (void)setDataSourceAsset:(NSString*)asset withKey:(NSString*)key{
+- (void)setDataSourceAsset:(NSString*)asset withKey:(NSString*)key overriddenDuration:(int) overriddenDuration{
     NSString* path = [[NSBundle mainBundle] pathForResource:asset ofType:nil];
-    return [self setDataSourceURL:[NSURL fileURLWithPath:path] withKey:key withHeaders: @{} withCache: false];
+    return [self setDataSourceURL:[NSURL fileURLWithPath:path] withKey:key withHeaders: @{} withCache: false overriddenDuration:overriddenDuration];
 }
 
-- (void)setDataSourceURL:(NSURL*)url withKey:(NSString*)key withHeaders:(NSDictionary*)headers withCache:(BOOL)useCache{
+- (void)setDataSourceURL:(NSURL*)url withKey:(NSString*)key withHeaders:(NSDictionary*)headers withCache:(BOOL)useCache overriddenDuration:(int) overriddenDuration{
     if (headers == [NSNull null]){
         headers = @{};
     }
@@ -268,6 +266,11 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
     } else {
         item = [AVPlayerItem playerItemWithURL:url];
     }
+    
+    if (@available(iOS 10.0, *) && overriddenDuration > 0) {
+        item.forwardPlaybackEndTime = CMTimeMake(overriddenDuration/1000, 1);
+    }
+    [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(playbackStalled:) name:AVPlayerItemPlaybackStalledNotification object:item ];
     
     return [self setDataSourcePlayerItem:item withKey:key];
 }
@@ -309,6 +312,16 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
     [self addObservers:item];
 }
 
+- (void)playbackStalled:(NSNotification *)notification {
+    if (_eventSink != nil) {
+        _eventSink([FlutterError
+                    errorWithCode:@"VideoError"
+                    message:@"Failed to load video: playback stalled"
+                    details:nil]);
+    }
+}
+
+
 - (void)observeValueForKeyPath:(NSString*)path
                       ofObject:(id)object
                         change:(NSDictionary*)change
@@ -319,7 +332,15 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
             for (NSValue* rangeValue in [object loadedTimeRanges]) {
                 CMTimeRange range = [rangeValue CMTimeRangeValue];
                 int64_t start = FLTCMTimeToMillis(range.start);
-                [values addObject:@[ @(start), @(start + FLTCMTimeToMillis(range.duration)) ]];
+                int64_t end = start + FLTCMTimeToMillis(range.duration);
+                if (!CMTIME_IS_INVALID(_player.currentItem.forwardPlaybackEndTime)) {
+                    int64_t endTime = FLTCMTimeToMillis(_player.currentItem.forwardPlaybackEndTime);
+                    if (end > endTime){
+                        end = endTime;
+                    }
+                }
+                
+                [values addObject:@[ @(start), @(end) ]];
             }
             _eventSink(@{@"event" : @"bufferingUpdate", @"values" : values, @"key" : _key});
         }
@@ -436,6 +457,10 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
         time =  [[_player currentItem] duration];
     } else {
         time =  [[[_player currentItem] asset] duration];
+    }
+    if (!CMTIME_IS_INVALID(_player.currentItem.forwardPlaybackEndTime)) {
+        NSLog(@"Using forwardplaybackedntime!");
+        time = [[_player currentItem] forwardPlaybackEndTime];
     }
     
     return FLTCMTimeToMillis(time);
@@ -805,7 +830,7 @@ NSMutableDictionary*  _artworkImageDict;
         [playerToRemoveListener.player removeTimeObserver: timeObserverId];
     }
     [_timeObserverIdDict removeAllObjects];
-  
+    
 }
 
 
@@ -839,7 +864,10 @@ NSMutableDictionary*  _artworkImageDict;
             NSString* uriArg = dataSource[@"uri"];
             NSString* key = dataSource[@"key"];
             NSDictionary* headers = dataSource[@"headers"];
-            
+            int overriddenDuration = 0;
+            if ([dataSource objectForKey:@"overriddenDuration"] != [NSNull null]){
+                overriddenDuration = [dataSource[@"overriddenDuration"] intValue];
+            }
             
             BOOL useCache = false;
             id useCacheObject = [dataSource objectForKey:@"useCache"];
@@ -858,9 +886,9 @@ NSMutableDictionary*  _artworkImageDict;
                 } else {
                     assetPath = [_registrar lookupKeyForAsset:assetArg];
                 }
-                [player setDataSourceAsset:assetPath withKey:key];
+                [player setDataSourceAsset:assetPath withKey:key overriddenDuration:overriddenDuration];
             } else if (uriArg) {
-                [player setDataSourceURL:[NSURL URLWithString:uriArg] withKey:key withHeaders:headers withCache: useCache];
+                [player setDataSourceURL:[NSURL URLWithString:uriArg] withKey:key withHeaders:headers withCache: useCache overriddenDuration:overriddenDuration];
             } else {
                 result(FlutterMethodNotImplemented);
             }