+++ lib/better_player.dart
@@ -2,16 +2,16 @@ library better_player;
 
 export 'src/configuration/better_player_cache_configuration.dart';
 export 'src/configuration/better_player_configuration.dart';
+export 'src/configuration/better_player_controls_configuration.dart';
 export 'src/configuration/better_player_data_source.dart';
 export 'src/configuration/better_player_data_source_type.dart';
 export 'src/configuration/better_player_event.dart';
 export 'src/configuration/better_player_event_type.dart';
 export 'src/configuration/better_player_notification_configuration.dart';
+export 'src/configuration/better_player_theme.dart';
 export 'src/configuration/better_player_translations.dart';
-export 'src/controls/better_player_controls_configuration.dart';
 export 'src/controls/better_player_overflow_menu_item.dart';
 export 'src/controls/better_player_progress_colors.dart';
-export 'src/controls/better_player_theme.dart';
 export 'src/core/better_player.dart';
 export 'src/core/better_player_controller.dart';
 export 'src/core/better_player_controller_provider.dart';