+++ lib/src/controls/better_player_theme.dart
@@ -1,1 +1,0 @@
-enum PlayerTheme { material, cupertino, custom }