+++ lib/src/core/better_player_controller.dart
@@ -9,6 +9,7 @@ import 'package:better_player/src/configuration/better_player_event.dart';
 import 'package:better_player/src/configuration/better_player_event_type.dart';
 import 'package:better_player/src/configuration/better_player_translations.dart';
 import 'package:better_player/src/core/better_player_controller_provider.dart';
+
 // Flutter imports:
 import 'package:better_player/src/core/better_player_utils.dart';
 import 'package:better_player/src/hls/better_player_hls_track.dart';
@@ -17,6 +18,7 @@ import 'package:better_player/src/subtitles/better_player_subtitle.dart';
 import 'package:better_player/src/subtitles/better_player_subtitles_factory.dart';
 import 'package:better_player/src/video_player/video_player.dart';
 import 'package:flutter/material.dart';
+
 // Package imports:
 import 'package:path_provider/path_provider.dart';
 
@@ -231,6 +233,7 @@ class BetterPlayerController extends ChangeNotifier {
               _betterPlayerDataSource?.notificationConfiguration?.imageUrl,
           notificationChannelName: _betterPlayerDataSource
               ?.notificationConfiguration?.notificationChannelName,
+          overriddenDuration: _betterPlayerDataSource.overriddenDuration,
         );
 
         break;
@@ -245,6 +248,7 @@ class BetterPlayerController extends ChangeNotifier {
               _betterPlayerDataSource?.notificationConfiguration?.imageUrl,
           notificationChannelName: _betterPlayerDataSource
               ?.notificationConfiguration?.notificationChannelName,
+          overriddenDuration: _betterPlayerDataSource.overriddenDuration,
         );
         break;
       case BetterPlayerDataSourceType.memory:
@@ -261,6 +265,7 @@ class BetterPlayerController extends ChangeNotifier {
                 _betterPlayerDataSource?.notificationConfiguration?.imageUrl,
             notificationChannelName: _betterPlayerDataSource
                 ?.notificationConfiguration?.notificationChannelName,
+            overriddenDuration: _betterPlayerDataSource.overriddenDuration,
           );
           _tempFiles.add(file);
         } else {